@@ -1,47 +1,49 @@
 {
 	"name": "sveltekit-tailwind-starter",
 	"version": "0.0.1",
-	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"test": "playwright test",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint .",
 		"test:unit": "vitest",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"test": "npm run test:unit -- --run && npm run test:e2e",
+		"test:e2e": "playwright test"
 	},
 	"devDependencies": {
-		"@inlang/sdk-js": "^0.11.0",
-		"@playwright/test": "^1.34.3",
-		"@poppanator/sveltekit-svg": "^5.0.0",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@tailwindcss/forms": "^0.5.3",
-		"@typescript-eslint/eslint-plugin": "^8.0.0",
-		"@typescript-eslint/parser": "^8.0.0",
-		"autoprefixer": "^10.4.14",
-		"eslint": "^9.0.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.30.0",
-		"postcss": "^8.4.24",
-		"postcss-load-config": "^4.0.1",
-		"prettier": "^3.0.0",
-		"prettier-plugin-svelte": "^3.0.0",
-		"prettier-plugin-tailwindcss": "^0.6.0",
-		"svelte": "^3.59.1",
-		"svelte-check": "^3.4.3",
-		"svelte-preprocess": "^5.0.4",
-		"tailwindcss": "^3.3.2",
-		"tslib": "^2.5.2",
-		"typescript": "^5.1.3",
-		"vite": "^5.0.0",
-		"vitest": "^0.34.0"
+		"@playwright/test": "1.48.2",
+		"@poppanator/sveltekit-svg": "5.0.0",
+		"@sveltejs/adapter-auto": "3.3.1",
+		"@sveltejs/kit": "2.8.1",
+		"@sveltejs/vite-plugin-svelte": "4.0.1",
+		"@tailwindcss/aspect-ratio": "0.4.2",
+		"@tailwindcss/container-queries": "0.1.1",
+		"@tailwindcss/forms": "0.5.9",
+		"@tailwindcss/typography": "0.5.15",
+		"@types/eslint": "9.6.1",
+		"autoprefixer": "10.4.20",
+		"eslint": "9.15.0",
+		"eslint-config-prettier": "9.1.0",
+		"eslint-plugin-svelte": "2.46.0",
+		"globals": "15.12.0",
+		"mode-watcher": "0.4.1",
+		"prettier": "3.3.3",
+		"prettier-plugin-svelte": "3.2.8",
+		"prettier-plugin-tailwindcss": "0.6.8",
+		"svelte": "5.2.2",
+		"svelte-check": "4.0.9",
+		"tailwindcss": "3.4.15",
+		"typescript": "5.6.3",
+		"typescript-eslint": "8.14.0",
+		"vite": "5.4.11",
+		"vitest": "2.1.5"
 	},
 	"dependencies": {
-		"@neoconfetti/svelte": "^1.0.0"
-	},
-	"type": "module"
+		"@fontsource-variable/inter": "5.1.0",
+		"@inlang/paraglide-sveltekit": "0.11.5"
+	}
 }
