@@ -1,23 +1,7 @@
-export const handle = async ({ event, resolve }) => {
-	let theme: string | null = null;
+import type { Handle } from '@sveltejs/kit';
 
-	const newTheme = event.url.searchParams.get('theme');
-	const cookieTheme = event.cookies.get('colortheme');
+import { i18n } from '$lib/i18n';
 
-	if (newTheme) {
-		theme = newTheme;
-	} else if (cookieTheme) {
-		theme = cookieTheme;
-	}
+const handleParaglide: Handle = i18n.handle();
 
-	if (theme) {
-		return await resolve(event, {
-			transformPageChunk: ({ html }) =>
-				html
-					.replace('data-theme=""', `data-theme="${theme}"`)
-					.replace('style="color-scheme: normal"', `style="color-scheme: ${theme}"`)
-		});
-	}
-
-	return await resolve(event);
-};
+export const handle: Handle = handleParaglide;
