@@ -1,9 +1,8 @@
 declare module '*.svg?component' {
-	import type { ComponentType, SvelteComponentTyped } from 'svelte';
+	import type { ComponentType, SvelteComponent } from 'svelte';
 	import type { SVGAttributes } from 'svelte/elements';
 
-	const content: ComponentType<SvelteComponentTyped<SVGAttributes<SVGSVGElement>>>;
-
+	const content: ComponentType<SvelteComponent<SVGAttributes<SVGSVGElement>>>;
 	export default content;
 }
 
