@@ -1,19 +1,16 @@
-import preprocess from 'svelte-preprocess';
 import adapter from '@sveltejs/adapter-auto';
-import { vitePreprocess } from '@sveltejs/kit/vite';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	// Consult https://github.com/sveltejs/svelte-preprocess
+	// Consult https://svelte.dev/docs/kit/integrations
 	// for more information about preprocessors
-	preprocess: [
-		vitePreprocess(),
-		preprocess({
-			postcss: true
-		})
-	],
+	preprocess: vitePreprocess(),
 
 	kit: {
+		// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
+		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+		// See https://svelte.dev/docs/kit/adapters for more information about adapters.
 		adapter: adapter(),
 		alias: {
 			$actions: 'src/lib/actions',
@@ -21,13 +18,6 @@ const config = {
 			$icons: 'src/lib/icons',
 			$stores: 'src/lib/stores',
 			$utils: 'src/lib/utils'
-		}
-	},
-
-	vitePlugin: {
-		inspector: {
-			holdMode: true,
-			toggleKeyCombo: 'meta-shift'
 		}
 	}
 };
